Separate off plot_parameters
